requirements.txt
Django==4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
scikit-learn==1.0.2
joblib==1.1.0